services/rag_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
langchain==0.1.9
langchain-community==0.0.24
langchain-ollama==0.0.1
faiss-cpu==1.7.4
python-multipart==0.0.9
httpx==0.26.0
tqdm==4.66.2
pypdf==4.0.1
python-pptx==0.6.23 